requirements.txt
langgraph 
langchain 
langchain-openai 
langchain-community
langchain-tavily
python-dotenv
pandas